Quick fix to introduce new IntegerLiteral into parser (needs proper handling)

+++ src/main/java/com/github/firmwehr/gentle/parser/tokens/IntegerLiteralToken.java
@@ -2,9 +2,11 @@ package com.github.firmwehr.gentle.parser.tokens;
 
 import com.github.firmwehr.gentle.source.SourceSpan;
 
+import java.math.BigInteger;
+
 public record IntegerLiteralToken(
 	SourceSpan sourceSpan,
-	int value
+	BigInteger value
 ) implements Token {
 	@Override
 	public String format() {
